+++ src/trx/game/cutseq.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <trx/game/cutseq/decoder.h>
+#include <trx/game/cutseq/pak.h>
+#include <trx/game/cutseq/playback.h>